+++ runtime/v1/project-model/src/main/java/com/walmartlabs/concord/project/InternalConstants.java
@@ -21,7 +21,6 @@ package com.walmartlabs.concord.project;
  */
 
 import com.walmartlabs.concord.sdk.Constants;
-import io.takari.bpm.model.ProcessDefinition;
 
 /**
  * Implementation-specific constants.